% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_numbers}
\alias{add_numbers}
\title{Add numbers function}
\usage{
add_numbers(num1, num2)
}
\arguments{
\item{num1}{Number to be added 1}

\item{num2}{Number to be added 2}
}
\value{
Returns the combined total of num1 and num2
}
\description{
Add numbers function
}
