% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_data_advanced}
\alias{import_data_advanced}
\title{Import data and create a comnbined tibble}
\usage{
import_data_advanced(dataType, mainDir)
}
\arguments{
\item{dataType}{The data file you wish to use}

\item{mainDir}{The main dir location}
}
\value{

}
\description{
Import data and create a comnbined tibble
}
