% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_user_info}
\alias{import_user_info}
\title{Get user_info_stuff}
\usage{
import_user_info(file_path)
}
\arguments{
\item{file_path}{Give me the filepath. You can use the herrrrrrre package.}
}
\value{

}
\description{
Get user_info_stuff
}
